requirements.txt
numpy==1.20.3
Pillow==9.4.0
torch==2.0.0
torchvision==0.15.1
wandb==0.14.0
